update: new logic functions, fix: wrong download size
